modarb_API/requirements.txt
fastapi
uvicorn[standard]
datasets
transformers 
accelerate 
evaluate  
bitsandbytes 
accelerate 
einops
Pillow
torch
torchvision

